pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agora-mem"
version = "0.1.0"
description = "Workflow-native session memory for AI agents and team handoffs."
authors = [
    { name = "thebnbrkr", email = "your-email@example.com" }
]
license = { text = "Apache-2.0" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=23.0",
    "aiosqlite>=0.19",
]

[project.optional-dependencies]
# Embedding + compression support
openai = ["openai>=1.0"]
gemini = ["google-generativeai>=0.5"]
# Cloud storage via Supabase
supabase = ["supabase>=2.0"]
# MCP server
mcp = ["mcp>=1.0"]
# Optional Agora integration (TracedMemoryNode)
agora = ["agora>=0.1.0"]
# Install everything
all = [
    "openai>=1.0",
    "google-generativeai>=0.5",
    "supabase>=2.0",
    "mcp>=1.0",
]

[project.scripts]
agora-mem-mcp = "agora_mem.mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["agora_mem*"]
